modify frontend struct
