--- docker-compose.yml
@@ -1,42 +1,0 @@
-version: "3"
-services:
-  client:
-    build:
-      context: ./frontend
-    networks:
-      - backend
-  server:
-    build:
-      context: ./backend
-    networks:
-      - backend
-    depends_on:
-      - mongo
-
-  nginx:
-    restart: always
-    build:
-      context: ./nginx
-    ports:
-      - "80:80"
-    networks:
-      - backend
-  mongo:
-    container_name: mongo
-    image: mongo
-    environment:
-      MONGO_INITDB_ROOT_USERNAME: root
-      MONGO_INITDB_ROOT_PASSWORD: root
-      MONGO_INITDB_DATABASE: abaotest
-    ports:
-      - "27017:27017"
-    networks:
-      - backend
-
-networks:
-  backend:
-    driver: bridge
-
-volumes:
-  data:
-    driver: local
