+++ frontend/src/css/Layout.scss
@@ -12,7 +12,7 @@
 
 /* header */
 .header {
-    width: calc(100%-2*$padding);
+ //   width: calc(100-2*$padding);
     background-color: $main-color;
     color: black;
     display: flex;
