+++ frontend/src/css/base/_reset.scss
@@ -141,3 +141,8 @@ button{
     outline: none;
   }
 }
+
+#root {
+  width: 100vw;
+  height: 100vh;
+}
