+++ frontend/src/pages/index.tsx
@@ -3,6 +3,6 @@ export { default as SetPet } from "./SetPet";
 export { default as PetCustom } from "./PetCustom";
 export { default as Account } from "./Account";
 export { default as NotFound } from "./404";
-export { default as Main } from './Main';
+export { default as Main } from './PlacePage';
 export { default as PlacePage } from "./PlacePage";
 export { default as PlaceInfoPage } from "./PlaceInfoPage";
