+++ frontend/src/routes/index.tsx
@@ -1,13 +1,11 @@
 import React from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import { Login, PlacePage, PlaceInfoPage, NotFound, Main } from "../pages";
-import { Introduction, PlacePage, PlaceInfoPage, NotFound, Main } from "../pages";
+import { Introduction, PlacePage, PlaceInfoPage, NotFound } from "../pages";
 
 const Router: React.FunctionComponent = () => (
 	<BrowserRouter>
 		<Switch>
-      <Route exact path="/" component={Introduction} />
-			<Route path="/main" component={Main} />
+			<Route exact path="/" component={Introduction} />
 			<Route path="/place" component={PlacePage} />
 			<Route path="/place/:id" component={PlaceInfoPage} />
 			<Route path="*" component={NotFound} />
