--- nginx/Dockerfile
@@ -1,5 +1,0 @@
-FROM nginx:latest
-
-COPY ./nginx.conf /etc/nginx/conf.d/default.conf
-
-EXPOSE 80
