--- nginx/nginx.conf
@@ -1,32 +1,0 @@
-upstream client {
-    server client:3000;
-}
-
-upstream server {
-    server server:3001;
-}
-
-server {
-    listen 80;
-
-    location / {        
-        proxy_pass http://client;
-    }
-
-    location /sockjs-node {
-        proxy_pass http://client;
-        proxy_http_version 1.1;
-        proxy_set_header Upgrade $http_upgrade;
-        proxy_set_header Connection "Upgrade";
-    }
-
-    location /api {
-        rewrite /api/(.*) /$1 break;
-        proxy_pass http://server;
-    }
-
-    location /docs {
-        proxy_pass http://server;
-    }
-
-}
